+++ Source/Cubiquity/Private/CubiquityTerrainVolume.cpp
@@ -61,11 +61,11 @@ FVector ACubiquityTerrainVolume::pickSurface(FVector localStartPosition, FVector
 
 void ACubiquityTerrainVolume::setVoxel(FVector position, const UCubiquityMaterialSet* materialSet)
 {
-	m_volume->setVoxel({ position.X, position.Y, position.Z }, *materialSet);
+	m_volume->setVoxel({ int32_t(position.X), int32_t(position.Y), int32_t(position.Z) }, *materialSet);
 }
 
 UCubiquityMaterialSet* ACubiquityTerrainVolume::getVoxel(FVector position) const
 {
-	const auto& voxel = m_volume->getVoxel({ position.X, position.Y, position.Z });
+	const auto& voxel = m_volume->getVoxel({ int32_t(position.X), int32_t(position.Y), int32_t(position.Z) });
 	return new UCubiquityMaterialSet(voxel);
 }
